requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
psycopg2-binary
requests
supabase
# RAG（with_rag=True 時需要）
langchain-text-splitters
langchain-openai
langchain-community
langchain-core
faiss-cpu
pypdf
docx2txt
python-pptx
unstructured
# google-generativeai  # 若要用 Gemini 請取消註解
# RAG／transcript：YouTube 字幕
youtube-transcript-api>=0.6.2